# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Furniskuy APIs",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/barang-keluars": {
            "get": {
                "tags": ["BarangKeluar"],
                "summary": "getBarangKeluarList",
                "description": "Get all BarangKeluars",
                "operationId": "75c59269ac55dd545c5b117ed8578861",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/BarangKeluar"
                                            }
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["BarangKeluar"],
                "summary": "createBarangKeluar",
                "description": "Create BarangKeluar",
                "operationId": "209c67d3bad752a9ba84da51182d2f91",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BarangKeluar"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/BarangKeluar"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/barang-keluars/{id}": {
            "get": {
                "tags": ["BarangKeluar"],
                "summary": "getBarangKeluarItem",
                "description": "Get BarangKeluar",
                "operationId": "fa031454d0103814c09a2b2f4c7aacce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of BarangKeluar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/BarangKeluar"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": ["BarangKeluar"],
                "summary": "updateBarangKeluar",
                "description": "Update BarangKeluar",
                "operationId": "5fa4dd2037aa558866162a0472dcdbbc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of BarangKeluar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BarangKeluar"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/BarangKeluar"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["BarangKeluar"],
                "summary": "deleteBarangKeluar",
                "description": "Delete BarangKeluar",
                "operationId": "76d8aaede2d4c2b66a48807042fed7de",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of BarangKeluar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "BarangKeluar": {
                "required": [],
                "properties": {
                    "tanggal": {
                        "description": "",
                        "type": "string",
                        "format": "date",
                        "readOnly": false,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    }
}